Added sync of last modified date from server and retry counter for deletions
The client sets the last modified date of a downloaded file/folder to the date
of the server side.

If a file is deleted locally, but could not be deleted on the server, it is skipped if
the remote operation failed 3 times. Also a local change on a write protected
file will be skipped after the third failed remote operation.

+++ CmisSync.Lib/Sync/SynchronizedFolder.cs
@@ -499,10 +499,15 @@ namespace CmisSync.Lib.Sync
 
                     // Delete the folder from the remote server.
                     Logger.Debug(String.Format("CMIS::DeleteTree({0})",remoteSubFolder.Path));
-                    remoteSubFolder.DeleteTree(true, null, true);
-
-                    // Delete the folder from database.
-                    database.RemoveFolder(localSubFolder);
+                    try{
+                        remoteSubFolder.DeleteTree(true, null, true);
+                        // Delete the folder from database.
+                        database.RemoveFolder(localSubFolder);
+                    }catch(Exception)
+                    {
+                        Logger.Info("Remote Folder could not be deleted: "+ remoteSubFolder.Path);
+                        // Just go on and try it the next time
+                    }
                 }
                 else
                 {
@@ -553,6 +558,7 @@ namespace CmisSync.Lib.Sync
                 // It sometimes happen on IBM P8 CMIS server, not sure why.
                 if (remoteDocument.ContentStreamFileName == null)
                 {
+                    //TODO Possibly the file content has been changed to 0, this case should be handled
                     Logger.Warn("Skipping download of '" + fileName + "' with null content stream in " + localFolder);
                     return true;
                 }
@@ -562,13 +568,12 @@ namespace CmisSync.Lib.Sync
                     remoteFiles.Add(fileName);
                 }
 
+                // Check if file extension is allowed
                 if (!Utils.WorthSyncing(fileName))
                 {
                     Logger.Info("Ignore the unworth syncing remote file: " + fileName);
                     return true;
                 }
-
-                // Check if file extension is allowed
 
                 bool success = true;
 
@@ -608,11 +613,24 @@ namespace CmisSync.Lib.Sync
                     {
                         if (database.ContainsFile(filePath))
                         {
-                            // File has been recently removed locally, so remove it from server too.
-                            Logger.Info("Removing locally deleted file on server: " + filePath);
-                            remoteDocument.DeleteAllVersions();
-                            // Remove it from database.
-                            database.RemoveFile(filePath);
+                            long retries = database.GetOperationRetryCounter(filePath, Database.OperationType.DELETE);
+                            if(retries < repoinfo.MaxDeletionRetries) {
+                                // File has been recently removed locally, so remove it from server too.
+                                Logger.Info("Removing locally deleted file on server: " + filePath);
+                                try{
+                                    remoteDocument.DeleteAllVersions();
+                                    // Remove it from database.
+                                    database.RemoveFile(filePath);
+                                    database.SetOperationRetryCounter(filePath, 0, Database.OperationType.DELETE);
+                                }catch(CmisBaseException ex)
+                                {
+                                    Logger.Warn("Could not delete remote file: ", ex);
+                                    database.SetOperationRetryCounter(filePath, retries+1, Database.OperationType.DELETE);
+                                    throw;
+                                }
+                            } else {
+                                Logger.Info(String.Format("Skipped deletion of remote file {0} because of too many failed retries ({1})", filePath, retries));
+                            }
                         }
                         else
                         {
@@ -631,6 +649,51 @@ namespace CmisSync.Lib.Sync
                 return success;
             }
 
+            private void SetLastModifiedDate(IDocument remoteDocument, string filepath, Dictionary<string, string[]> metadata)
+            {
+                if (remoteDocument.LastModificationDate != null)
+                {
+                    File.SetLastWriteTimeUtc(filepath, (DateTime)remoteDocument.LastModificationDate);
+                }
+                else
+                {
+                    string[] cmisModDate;
+                    if (metadata.TryGetValue("cmis:lastModificationDate", out cmisModDate) && cmisModDate.Length == 3)
+                    {
+                        try
+                        {
+                            DateTime modDate = DateTime.Parse(cmisModDate[2]);
+                            File.SetLastWriteTimeUtc(filepath, modDate);
+                        }
+                        catch (Exception)
+                        {
+                        }
+                    }
+                }
+            }
+
+            private void SetLastModifiedDate(IFolder remoteFolder, string filepath, Dictionary<string, string[]> metadata)
+            {
+                if (remoteFolder.LastModificationDate != null)
+                {
+                    File.SetLastWriteTimeUtc(filepath, (DateTime)remoteFolder.LastModificationDate);
+                }
+                else
+                {
+                    string[] cmisModDate;
+                    if (metadata.TryGetValue("cmis:lastModificationDate", out cmisModDate) && cmisModDate.Length == 3)
+                    {
+                        try
+                        {
+                            DateTime modDate = DateTime.Parse(cmisModDate[2]);
+                            File.SetLastWriteTimeUtc(filepath, modDate);
+                        }
+                        catch (Exception)
+                        {
+                        }
+                    }
+                }
+            }
 
             /// <summary>
             /// Download a single file from the CMIS server.
@@ -653,7 +716,11 @@ namespace CmisSync.Lib.Sync
                         DotCMIS.Data.IContentStream contentStream = null;
                         string filepath = Path.Combine(localFolder, fileName);
                         string tmpfilepath = filepath + ".sync";
-
+                        if(database.GetOperationRetryCounter(filepath,Database.OperationType.DOWNLOAD) > repoinfo.MaxDownloadRetries)
+                        {
+                            Logger.Info(String.Format("Skipping download of file {0} because of too many failed ({1}) downloads",database.GetOperationRetryCounter(filepath,Database.OperationType.DOWNLOAD)));
+                            return true;
+                        }
                         // If there was previously a directory with this name, delete it.
                         // TODO warn if local changes inside the folder.
                         if (Directory.Exists(filepath))
@@ -765,6 +832,7 @@ namespace CmisSync.Lib.Sync
                         catch (System.IO.DirectoryNotFoundException ex)
                         {
                             Logger.Warn(String.Format("Download failed because of a missing folder in the file path: {0}" , ex.Message ));
+                            success = false;
                         }
                         catch (Exception ex)
                         {
@@ -772,6 +840,11 @@ namespace CmisSync.Lib.Sync
                             success = false;
                             Logger.Debug("Removing temp download file: "+ tmpfilepath);
                             File.Delete(tmpfilepath);
+                            success = false;
+                            if(ex is CmisBaseException)
+                            {
+                                database.SetOperationRetryCounter(filepath,database.GetOperationRetryCounter(filepath,Database.OperationType.DOWNLOAD)+1,Database.OperationType.DOWNLOAD);
+                            }
                         }
 
                         if (success)
@@ -811,7 +884,7 @@ namespace CmisSync.Lib.Sync
                                     File.Move(filepath, newFilePath);
                                     Logger.Debug(String.Format("Moving temporary local download file {0} to target file {1}", tmpfilepath, filepath));
                                     File.Move(tmpfilepath, filepath);
-
+                                    SetLastModifiedDate(remoteDocument, filepath, metadata);
                                     repo.OnConflictResolved();
                                     // TODO move to OS-dependant layer
                                     //System.Windows.Forms.MessageBox.Show("Someone modified a file at the same time as you: " + filePath
@@ -824,12 +897,15 @@ namespace CmisSync.Lib.Sync
                                     File.Delete(filepath);
                                     Logger.Debug(String.Format("Moving temporary local download file {0} to target file {1}", tmpfilepath, filepath));
                                     File.Move(tmpfilepath, filepath);
+                                    SetLastModifiedDate(remoteDocument, filepath, metadata);
+                                    
                                 }
                             }
                             else
                             {
                                 Logger.Debug(String.Format("Moving temporary local download file {0} to target file {1}", tmpfilepath, filepath));
                                 File.Move(tmpfilepath, filepath);
+                                SetLastModifiedDate(remoteDocument, filepath, metadata);
                             }
 
                             // Create database entry for this file.
@@ -950,6 +1026,7 @@ namespace CmisSync.Lib.Sync
                             Dictionary<string, object> properties = new Dictionary<string, object>();
                             properties.Add(PropertyIds.Name, fileName);
                             properties.Add(PropertyIds.ObjectTypeId, "cmis:document");
+                            properties.Add(PropertyIds.CreationDate, ((long)(File.GetCreationTimeUtc(filePath) - new DateTime(1970, 1, 1)).TotalMilliseconds).ToString());
 
                             // Prepare content stream
                             using (Stream file = File.OpenRead(filePath))
@@ -1032,6 +1109,7 @@ namespace CmisSync.Lib.Sync
 
                             // Create database entry for this file.
                             database.AddFile(filePath, remoteDocument.Id, remoteDocument.LastModificationDate, metadata, filehash);
+                            SetLastModifiedDate(remoteDocument, filePath, metadata);
                         }
                         return success;
                     }
@@ -1056,6 +1134,8 @@ namespace CmisSync.Lib.Sync
                 Dictionary<string, object> properties = new Dictionary<string, object>();
                 properties.Add(PropertyIds.Name, Path.GetFileName(localFolder));
                 properties.Add(PropertyIds.ObjectTypeId, "cmis:folder");
+                properties.Add(PropertyIds.CreationDate, "");
+                properties.Add(PropertyIds.LastModificationDate,"");
                 IFolder folder = null;
                 try
                 {
@@ -1086,7 +1166,7 @@ namespace CmisSync.Lib.Sync
                 // Create database entry for this folder
                 // TODO Add metadata
                 database.AddFolder(localFolder, folder.Id, folder.LastModificationDate);
-
+                SetLastModifiedDate(folder,localFolder, FetchMetadata(folder));
                 bool success = true;
                 try
                 {
@@ -1355,15 +1435,15 @@ namespace CmisSync.Lib.Sync
             /// Retrieve the CMIS metadata of a document.
             /// </summary>
             /// <returns>a dictionary in which each key is a type id and each value is a couple indicating the mode ("readonly" or "ReadWrite") and the value itself.</returns>
-            private Dictionary<string, string[]> FetchMetadata(IDocument document)
+            private Dictionary<string, string[]> FetchMetadata(ICmisObject o)
             {
                 Dictionary<string, string[]> metadata = new Dictionary<string, string[]>();
 
-                IObjectType typeDef = session.GetTypeDefinition(document.ObjectType.Id/*"cmis:document" not Name FullName*/); // TODO cache
+                IObjectType typeDef = session.GetTypeDefinition(o.ObjectType.Id/*"cmis:document" not Name FullName*/); // TODO cache
                 IList<IPropertyDefinition> propertyDefs = typeDef.PropertyDefinitions;
 
                 // Get metadata.
-                foreach (IProperty property in document.Properties)
+                foreach (IProperty property in o.Properties)
                 {
                     // Mode
                     string mode = "readonly";
